requirements.txt
openai>=1.0.0
google-api-python-client>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.0.0
